requirements.txt
asgiref==3.4.1
Django==4.0.1
django-cors-headers==3.11.0
djangorestframework==3.13.1
Pillow==9.0.0
pytesseract==0.3.8
pytz==2021.3
sqlparse==0.4.2
tesseract==0.1.3
